+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/consts/consts.go
@@ -32,7 +32,8 @@ const (
 	NameOfNetworkToStartEngineAndLogServiceContainersIn = "bridge"
 	HttpApplicationProtocol                             = "http"
 
-	GitHubAuthStorageDirPath = "/kurtosis-data/github-auth/"
+	GitHubAuthStorageDirPath   = "/kurtosis-data/github-auth/"
+	DockerConfigStorageDirPath = "/root/.docker/"
 
 	EmptyApplicationURL = ""
 )